+++ frontend/dashboard/src/pages/Account/Accounts.tsx
@@ -1,0 +1,6 @@
+import React from 'react';
+
+export default function Accounts({ client }: any) {
+  console.log('ACCOUNTS ', client);
+  return <div>Accounts</div>;
+}
